+++ cmd/wcbot/main.go
@@ -6,11 +6,11 @@
 package main
 
 import (
+	"fmt"
 	"os"
 	"os/signal"
 
 	"github.com/BigOokie/skywire-wing-commander/internal/telegrambot"
-	"github.com/BigOokie/skywire-wing-commander/internal/utils"
 	"github.com/BigOokie/skywire-wing-commander/internal/wcconst"
 	log "github.com/sirupsen/logrus"
 )
@@ -18,7 +18,7 @@ import (
 var wc wcBotApp
 
 func main() {
-	// Setup and initalise application logging
+	// Setup and initialise application logging
 	wc.initLogging()
 
 	// Parse and handle known command line flags
@@ -33,12 +33,12 @@ func main() {
 	}
 
 	// Check and setup application instance control. Only allow a single instance to run
-	appInstance := utils.InitAppInstance(wcconst.AppInstanceID)
-	defer appInstance.TryUnlock()
+	//appInstance := utils.InitAppInstance(wcconst.AppInstanceID)
+	//defer utils.ReleaseAppInstance(appInstance)
 
-	// Setup OS Notification for Interupt or Kill signal - to cleanly terminate the app
+	// Setup OS Notification for Interrupt or Kill signal - to cleanly terminate the app
 	osSignal := make(chan os.Signal, 1)
-	signal.Notify(osSignal, os.Interrupt, os.Kill)
+	signal.Notify(osSignal, os.Interrupt)
 
 	log.Infoln("Skywire Wing Commander Telegram Bot - Starting.")
 	defer log.Infoln("Skywire Wing Commander Telegram Bot - Stopped.")
@@ -51,16 +51,23 @@ func main() {
 		return
 	}
 
+	var startmsg string
+	// Check to see if we are starting because of an upgrade.
+	if wc.cmdFlags.upgradecompleted {
+		startmsg = fmt.Sprintf("*Successfully restarted after upgrade to %s*", wcconst.BotVersion)
+	} else {
+		startmsg = fmt.Sprintf("*Started: %s*", wcconst.BotAppVersion)
+	}
+	log.Debug(startmsg)
+	err = bot.SendNewMessage("markdown", startmsg)
+	if err != nil {
+		log.Fatalf("Failed to create Telegram updates channel: %v", err)
+	}
+
 	log.Infoln("Starting Bot instance.")
 	go bot.Start()
 
 	// Wait for the app to be signaled to terminate
-	select {
-	case signal := <-osSignal:
-		if signal == os.Interrupt {
-			log.Debugln(wcconst.MsgOSInteruptSig)
-		} else if signal == os.Kill {
-			log.Debugln(wcconst.MsgOSKillSig)
-		}
-	}
+	signal := <-osSignal
+	log.Debugln(wcconst.MsgOSInteruptSig, signal)
 }