+++ internal/skymgrmon/skymgrmon.go
@@ -33,6 +33,8 @@ type SkyManagerMonitor struct {
 	connectedNodes       skynode.NodeInfoMap
 	discConnNodeCount    int
 	m                    sync.Mutex
+	updateStarted        bool
+	updateMsgChan        chan string
 }
 
 // SetCancelFunc is a thread-safe function for setting the cancelFunc
@@ -61,6 +63,22 @@ func (smm *SkyManagerMonitor) DoCancelFunc() {
 	}
 }
 
+// GetUpdateStarted is a thread-safe function for checking if the
+// updateStarted flag has been set
+func (smm *SkyManagerMonitor) GetUpdateStarted() bool {
+	smm.m.Lock()
+	defer smm.m.Unlock()
+	return smm.updateStarted
+}
+
+// SetUpdateStarted is a thread-safe function for setting the
+// updateStarted flag has been set
+func (smm *SkyManagerMonitor) SetUpdateStarted(flag bool) {
+	smm.m.Lock()
+	defer smm.m.Unlock()
+	smm.updateStarted = flag
+}
+
 // NewMonitor creates a SkyManagerMonitor which will monitor the provided managerip.
 func NewMonitor(manageraddress, discoveryaddress string) *SkyManagerMonitor {
 	return &SkyManagerMonitor{
@@ -70,16 +88,19 @@ func NewMonitor(manageraddress, discoveryaddress string) *SkyManagerMonitor {
 		monitorStatusMsgChan: nil,
 		connectedNodes:       make(skynode.NodeInfoMap),
 		discConnNodeCount:    0,
+		updateStarted:        false,
+		updateMsgChan:        nil,
 	}
 }
 
 // RunManagerMonitor starts the SkyManagerMonitor monitoring of the local Manager Node.
 // If `ctx` is not nil, the monitor will listen to ctx.Done() and stop monitoring
-// when it recieves the signal.
-func (smm *SkyManagerMonitor) RunManagerMonitor(runctx context.Context, statusMsgChan chan<- string, pollInt time.Duration) {
+// when it receives the signal.
+func (smm *SkyManagerMonitor) RunManagerMonitor(runctx context.Context, doCancelFunc func(), statusMsgChan chan<- string, pollInt time.Duration) {
 	log.Debugf("SkyManagerMonitor::RunManagerMonitor: Start (Interval: %v)", pollInt)
 	defer log.Debugln("SkyManagerMonitor::RunManagerMonitor: End")
 
+	smm.SetCancelFunc(doCancelFunc)
 	smm.monitorStatusMsgChan = statusMsgChan
 
 	ticker := time.NewTicker(pollInt)
@@ -102,24 +123,26 @@ func (smm *SkyManagerMonitor) RunManagerMonitor(runctx context.Context, statusMs
 	}
 }
 
-// StopManagerMonitor starts the SkyManagerMonitor monitoring of the local Manager Node.
+// StopManagerMonitor stops the SkyManagerMonitor monitoring of the local Manager Node.
 // If `ctx` is not nil, the monitor will listen to ctx.Done() and stop monitoring
-// when it recieves the signal.
+// when it receives the signal.
 func (smm *SkyManagerMonitor) StopManagerMonitor() {
-	log.Debugln("SkyManagerMonitor::SkyManagerMonitor: Start")
-	defer log.Debugln("SkyManagerMonitor::SkyManagerMonitor: End")
+	log.Debugln("SkyManagerMonitor::StopManagerMonitor: Start")
+	defer log.Debugln("SkyManagerMonitor::StopManagerMonitor: End")
+
 	if smm.IsRunning() {
 		smm.DoCancelFunc()
 		smm.SetCancelFunc(nil)
 		close(smm.monitorStatusMsgChan)
 		smm.monitorStatusMsgChan = nil
+		log.Debug(wcconst.MsgMonitorStopped)
 	}
 }
 
 /*
 // RunDiscoveryMonitor starts the SkyManagerMonitor monitoring of the Skywire Discovery Node.
 // If `ctx` is not nil, the monitor will listen to ctx.Done() and stop monitoring
-// when it recieves the signal.
+// when it receives the signal.
 func (smm *SkyManagerMonitor) RunDiscoveryMonitor(runctx context.Context, statusMsgChan chan<- string, pollInt time.Duration) {
 	log.Debugf("SkyManagerMonitor::RunDiscoveryMonitor: Start (Interval: %v)", pollInt)
 	defer log.Debugln("SkyManagerMonitor::RunDiscoveryMonitor: End")
@@ -172,7 +195,7 @@ func (smm *SkyManagerMonitor) ConnectedDiscNodeCount() (int, error) {
 		defer smm.m.Unlock()
 
 		// Compare the list of Nodes connected to the Discovery Node (disccns) against the
-		// current list of locally conected nodes.
+		// current list of locally connected nodes.
 		// If our local Nodes are not listed as connected to the Discovery Node we need to raise an alert
 		discNodeMap := skynode.NodeInfoSliceToMap(discNodes)
 		for _, v := range smm.connectedNodes {
@@ -197,6 +220,7 @@ func (smm *SkyManagerMonitor) IsRunning() bool {
 	return smm.GetCancelFunc() != nil
 }
 
+/*
 // getAllNodesStr requests the list of connected Nodes from the Manager and returns the raw JSON response as a string
 func (smm *SkyManagerMonitor) getAllNodesStr() string {
 	var respstr string
@@ -217,6 +241,7 @@ func (smm *SkyManagerMonitor) getAllNodesStr() string {
 	}
 	return respstr
 }
+*/
 
 // getAllNodesList requests the list of connected Nodes from the Manager and returns an array (slice) of connectedNode
 func getAllNodesList(managerAddr string) (cns skynode.NodeInfoSlice, err error) {
@@ -307,9 +332,9 @@ func (smm *SkyManagerMonitor) maintainConnectedNodesList(newcns skynode.NodeInfo
 			}
 		}
 	}
-	return
 }
 
+/*
 // checkNodeDiscoveryConnection is responsible for checking the list of Nodes currently connected to the local Manager
 // against the list of Nodes reported as connected to the Skywire Discovery Node. If our local Nodes are not reported
 // as connected to the Discovery Node, we need to raise an alert using the provided statusMsgChan
@@ -332,7 +357,7 @@ func (smm *SkyManagerMonitor) checkNodeDiscoveryConnection(disccns skynode.NodeI
 	discConnNodeCount := 0
 
 	// Compare the list of Nodes connected to the Discovery Node (disccns) against the
-	// current list of locally conected nodes.
+	// current list of locally connected nodes.
 	// If our local Nodes are not listed as connected to the Discovery Node we need to raise an alert
 	discNodeMap := skynode.NodeInfoSliceToMap(disccns)
 	for _, v := range smm.connectedNodes {
@@ -355,6 +380,7 @@ func (smm *SkyManagerMonitor) checkNodeDiscoveryConnection(disccns skynode.NodeI
 	//statusMsgChan <- msg
 	return
 }
+*/
 
 // GetConnectedNodeCount will return the count of Nodes within the connectedNodes structure
 // If the structure is nil (not yet assigned), 0 will be returned